chore: move requests stub
