+++ tests/requests_stub.py
